ArabicMoM-GraphQL/requirements.txt
streamlit==1.28.0
flask==2.3.3
flask-graphql==2.0.1
graphene==2.1.9
pandas==2.1.1
python-dotenv==1.0.0
watchdog==3.0.0
